% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixCausalMed.R
\name{MixCausalMed}
\alias{MixCausalMed}
\title{Integration of decision trees with mixture mediation model}
\usage{
MixCausalMed(data, K)
}
\arguments{
\item{data}{The data frame contains treatment variable, confounding variables, mediator variable and outcome variable.}

\item{K}{The number of mixing components}
}
\description{
Integration of decision trees with mixture mediation model
}
\examples{
R code examples of how to use your function

}
\keyword{Model}
\keyword{estimation}
\keyword{for}
\keyword{mediation}
\keyword{mixture}
\keyword{model}
\keyword{of}
